+++ einsum-derive/src/args.rs
@@ -1,0 +1,22 @@
+use proc_macro2::TokenStream as TokenStream2;
+use proc_macro_error::{abort_call_site, ResultExt};
+use syn::parse::Parser;
+
+pub fn parse(input: TokenStream2) -> (String, Vec<syn::Expr>) {
+    let parser = syn::punctuated::Punctuated::<syn::Expr, syn::Token![,]>::parse_terminated;
+    let args = parser
+        .parse2(input)
+        .expect_or_abort("Invalid input for einsum!");
+    let mut iter = args.into_iter();
+    let subscripts = if let Some(syn::Expr::Lit(syn::ExprLit {
+        lit: syn::Lit::Str(lit),
+        attrs: _,
+    })) = iter.next()
+    {
+        lit.value()
+    } else {
+        abort_call_site!("einsum! must start with subscript string literal")
+    };
+    let args = iter.collect::<Vec<_>>();
+    (subscripts, args)
+}
